refactor: split repository support helpers
